#2: Source files renamed and makefile updated
