+++ Statistics.cpp
@@ -30,7 +30,7 @@
 #include <limits.h>
 #include <cmath>
 
-#include "CPU_Statistics.h"
+#include "Statistics.h"
 
 //--------------------------------------------------------------------------------------------------------------------//
 //--------------------------------------------------- Definitions ----------------------------------------------------//
