+++ Statistics.h
@@ -29,7 +29,7 @@
 
 
 #include "Parameters.h"
-#include "CPU_Population.h"
+#include "Population.h"
 #include "GlobalKnapsackData.h"
 
 /**
